valentines_deployment/valentines_docker/requirements.txt
requests
fastapi
uvicorn
mangum
httpx
boto3
python-multipart
pydantic
typing
llama-cpp-python
accelerate